refactor: RecordingControls and inline more stuff
