+++ apps/app/src/routes/(config)/+layout.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import NavItems from '$lib/components/NavItems.svelte';
-	import SelectTransformationOrCancelRecordingSessionOrEndRecordingSessionButton from '$lib/components/RecordingControlsOrCancelRecordingSessionOrEndRecordingSessionButton.svelte';
+	import RecordingControls from '$lib/components/RecordingControls.svelte';
 	import WhisperingButton from '$lib/components/WhisperingButton.svelte';
 	import { getRecorderFromContext } from '$lib/query/singletons/recorder';
 	import { cn } from '$lib/utils.js';
@@ -27,7 +27,7 @@
 			<span class="text-lg font-bold">whispering</span>
 		</WhisperingButton>
 	</div>
-	<SelectTransformationOrCancelRecordingSessionOrEndRecordingSessionButton />
+	<RecordingControls />
 	<WhisperingButton
 		tooltipContent="Toggle recording"
 		onclick={recorder.toggleRecording}
